.github/scripts/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
applicationinsights==0.11.10
async-timeout==4.0.2
attrs==21.4.0
azure-cognitiveservices-speech==1.14.0
azure-core==1.23.0
azure-data-tables==12.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
dateparser==1.1.0
Deprecated==1.2.13
frozenlist==1.2.0
idna==3.3
isodate==0.6.1
lxml==4.8.0
msrest==0.6.21
multidict==5.2.0
oauthlib==3.2.0
packaging==21.3
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.7
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.1
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
websockets==9.1
wrapt==1.13.3
yarl==1.7.2
